feat: Add user ID validation to service layer for project and task access
-Updated the service layer to include business logic for validating user
IDs.
-Ensured that only authorized users can access information about
projects and tasks.
-Enhanced security by integrating ownership checks before providing data
to the user.

+++ internal/services/project_service.go
@@ -14,28 +14,37 @@ func NewProjectService(projectRepo *database.ProjectRepository) *ProjectService 
 	return &ProjectService{ProjectRepo: projectRepo}
 }
 
-func (p *ProjectService) CreateProject(project *models.Project) error {
+func (p *ProjectService) CreateProject(project *models.Project, userID uint) error {
 	if project.Name == "" {
 		return errors.New("project name can not be empty")
 	}
-	if project.UserID == 0 {
+	if userID == 0 {
 		return errors.New("project must associated with a user")
 	}
+	project.UserID = userID
 	return p.ProjectRepo.CreateProject(project)
 }
 
-func (p *ProjectService) GetProjectById(id uint) (*models.Project, error) {
+func (p *ProjectService) GetProjectById(userID, id uint) (*models.Project, error) {
 	project, err := p.ProjectRepo.GetProjectById(id)
 	if err != nil {
 		return nil, err
 	}
+	isAssigned, err := p.ProjectRepo.IsUserAssignedToProject(userID, id)
+	if err != nil {
+		return nil, err
+	}
+	if !isAssigned {
+		return nil, errors.New("project is not assigned to this user")
+	}
 	return project, nil
 }
 
-func (p *ProjectService) UpdateProject(project *models.Project) error {
+func (p *ProjectService) UpdateProject(project *models.Project, userID uint) error {
 	if project.Name == "" {
 		return errors.New("project name can not be empty")
 	}
+
 	projectUpdated, err := p.ProjectRepo.GetProjectById(project.ID)
 	if err != nil {
 		return err
@@ -43,16 +52,30 @@ func (p *ProjectService) UpdateProject(project *models.Project) error {
 	if projectUpdated == nil {
 		return errors.New("project doesn't exist")
 	}
+	isAssigned, err := p.ProjectRepo.IsUserAssignedToProject(userID, project.ID)
+	if err != nil {
+		return err
+	}
+	if !isAssigned {
+		return errors.New("project is not assigned to this user")
+	}
 	return p.ProjectRepo.UpdateProject(project)
 }
 
-func (p *ProjectService) DeleteProjectById(id uint) error {
+func (p *ProjectService) DeleteProjectById(userID, id uint) error {
 	project, err := p.ProjectRepo.GetProjectById(id)
 	if err != nil {
 		return err
 	}
 	if project == nil {
 		return errors.New("project doesn't exist")
+	}
+	isAssigned, err := p.ProjectRepo.IsUserAssignedToProject(userID, project.ID)
+	if err != nil {
+		return err
+	}
+	if !isAssigned {
+		return errors.New("project is not assigned to this user")
 	}
 	return p.ProjectRepo.DeleteProjectById(id)
 }
@@ -63,4 +86,8 @@ func (p *ProjectService) GetAllProjects() ([]*models.Project, error) {
 
 func (p *ProjectService) GetAllProjectByUserId(userId uint) ([]*models.Project, error) {
 	return p.ProjectRepo.GetAllProjectByUserId(userId)
+}
+
+func (p *ProjectService) IsUserAssignedToProject(projectId uint, userId uint) (bool, error) {
+	return p.ProjectRepo.IsUserAssignedToProject(projectId, userId)
 }
